data_science/cloud_functions/ProcessBattlesGSheets/requirements.txt
google-cloud-storage==1.26.0
requests==2.23.0
google-api-python-client==1.8.0